Feat: User Policies & Article Policies

+++ app/Http/Controllers/UserController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
+use App\Http\Requests\UserRequest;
 use App\Models\User;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
@@ -8,19 +9,34 @@ class UserController extends Controller
 {
 
 use AuthorizesRequests;
-    public function destroy($id) {
-    
-     $user = User::findOrFail($id);
-      $user->delete();
+    public function index() {
+      $this->authorize('viewAny', User::class);
+      $users = User::latest()->paginate(20);
+      return response()->json($users);
+    }
+    public function show(User $user) {
+      $this->authorize('view', $user);
+      return response()->json($user);
+  }
 
-    return response()->json(
-  ['message' => 'Gebruiker is verwijderd.']);
+    public function store(UserRequest $request) {
+      $this->authorize('create', User::class);
+      $user = User::create($request->validated());
+      return response()->json($user, 201);
     }
 
-    public function get($id) {
-        
-        $this->authorize('get', $id);
-        $user = User::findOrFail($id);
-        return response()->json($user);
+    public function update(UserRequest $request, User $user) {
+      $this->authorize('update', $user);
+      $user->update($request->validated());
+      return response()->json($user);
+    }
+
+    public function destroy(User $user) {
+      $this->authorize('delete', $user);
+      $user->delete();
+      return response()->json([
+      'message' => 'Gebruiker is verwijderd.'
+    ]);
     }
 }
+
